feat: enhancing channels (#28)(#30)(#31)(#25)(#24)(#23)

+++ Infrastructure/Services/Channels/ChannelService.cs
@@ -1,6 +1,10 @@
 ﻿using Application.Common.Interfaces.Channels;
 using Application.DTOs.Channels;
+using Application.Exceptions.Channels;
 using Domain.Entities;
+using Infrastructure.Channels.Tags;
+using Mapster;
+using MapsterMapper;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using System;
@@ -14,29 +18,38 @@ namespace Infrastructure.Services.Channels
     {
         private readonly AppDbContext _db;
         private readonly ILogger<ChannelService> _logger;
+
         public ChannelService(AppDbContext db, ILogger<ChannelService> logger)
         {
             _db = db;
             _logger = logger;
         }
 
-        public async Task<ChannelDto> CreateChannelAsync(ChannelCreateDto dto)
+        public async Task<ChannelDto> CreateChannelAsync(ChannelCreateDto dto, Guid ownerId)
         {
+            var publicId = await GeneratePublicIdAsync();
+            var entity = new Channel(dto.Name, dto.Description,
+                ownerId, publicId,
+                dto.IsPublic,
+                dto.IsPublicName,
+                dto.IsVisible);
 
-            var entity = new Channel(dto.Name, dto.OwnerId);
+            entity.AddMember(ownerId, ChannelMember.ChannelRole.Owner);
+            if (dto.Tags?.Any() == true)
+            {
+                var tags = await ResolveTagsAsync(dto.Tags);
 
+                foreach (var tag in tags)
+                {
+                    entity.AddTag(tag);
+                }
+            }
             _db.Channels.Add(entity);
             await _db.SaveChangesAsync();
 
             _logger.LogInformation("Channel created {ChannelId} by {OwnerId}", entity.Id, entity.OwnerId);
 
-            return new ChannelDto
-            {
-                Id = entity.Id,
-                Name = entity.Name,
-                OwnerId = entity.OwnerId,
-                CreatedAt = entity.CreatedAt
-            };
+            return entity.Adapt<ChannelDto>();
         }
 
         public async Task<ChannelDto> GetChannelAsync(Guid id)
@@ -50,24 +63,20 @@ namespace Infrastructure.Services.Channels
                 return null;
             }
 
-            return new ChannelDto
-            {
-                Id = entity.Id,
-                Name = entity.Name,
-                OwnerId = entity.OwnerId,
-                CreatedAt = entity.CreatedAt
-            };
+            return entity.Adapt<ChannelDto>();
         }
 
         public async Task<IEnumerable<ChannelDto>> GetAllChannelsAsync()
         {
             var list = await _db.Channels
+                .Where(c => c.IsVisible)
                 .Select(c => new ChannelDto
                 {
                     Id = c.Id,
                     Name = c.Name,
-                    OwnerId = c.OwnerId,
-                    CreatedAt = c.CreatedAt
+                    Description = c.Description,
+                    CreatedAt = c.CreatedAt,
+                    IsPublic = c.IsPublic,
                 })
                 .ToListAsync();
 
@@ -76,36 +85,76 @@ namespace Infrastructure.Services.Channels
             return list;
         }
 
-        public async Task<bool> UpdateChannelAsync(Guid id, ChannelUpdateDto dto)
+        public async Task<bool> UpdateChannelAsync(Guid id, ChannelUpdateDto dto,
+            Guid ownerId)
         {
-            var entity = await _db.Channels.FirstOrDefaultAsync(c => c.Id == id);
+            var entity = await _db.Channels
+                .Include(c => c.Tags)
+                .FirstOrDefaultAsync(c => c.Id == id);
 
             if (entity == null)
-            {
-                _logger.LogWarning("Update failed. Channel not found {ChannelId}", id);
                 return false;
+
+            if (entity.OwnerId != ownerId) throw new MemberNotAuthorizedException();
+
+            entity.Update(dto.Name, dto.Description,
+                dto.IsPublic,
+                dto.IsPublicName, dto.IsVisible);
+
+            if (dto.Tags != null)
+            {
+                var tags = await ResolveTagsAsync(dto.Tags);
+                entity.ReplaceTags(tags);
             }
 
-            var nameField = typeof(Channel)
-                .GetProperty("Name", BindingFlags.Instance | BindingFlags.NonPublic | BindingFlags.Public);
-
-            nameField.SetValue(entity, dto.Name);
-
             await _db.SaveChangesAsync();
-
-            _logger.LogInformation("Channel updated {ChannelId}", id);
             return true;
         }
-        public async Task<string> GetChannelByName(string name)
+
+        public async Task<IEnumerable<ChannelDto>> GetChannelsForUserAsync(Guid userId)
         {
-            var channel = await _db.Channels.FirstOrDefaultAsync
-                (c => c.Name == name);
-
-            return channel is null ? "Channel not found" : channel.Id.ToString();
-
-
+            return await _db.ChannelMembers
+                .Where(m => m.UserId == userId)
+                .Join(_db.Channels,
+                    m => m.ChannelId,
+                    c => c.Id,
+                    (m, c) => c)
+                .Select(c => c.Adapt<ChannelDto>())
+                .ToListAsync();
         }
-        public async Task<bool> DeleteChannelAsync(Guid id)
+
+        public async Task<IEnumerable<ChannelDto>> SearchChannelsByTags(IEnumerable<string> tags)
+        {
+            var normalized = tags
+                .Select(t => t.ToLowerInvariant())
+                .ToList();
+
+            return await _db.Channels
+                .Where(c =>
+                    c.Tags.Any(ct =>
+                        normalized.Contains(ct.Tag.NormalizedName)))
+                .Select(c => c.Adapt<ChannelDto>())
+                .ToListAsync();
+        }
+
+        public async Task<IEnumerable<ChannelDto?>> SearchByNameAsync(string channelName)
+        {
+            var channel = await _db.Channels
+                .Where(c => c.Name.StartsWith(channelName)
+                && c.IsVisible)
+                .ToListAsync();
+            return channel.Adapt<IEnumerable<ChannelDto>>();
+        }
+
+        public async Task<ChannelDto?> SearchByPublicIdAsync(int publicId)
+        {
+            var entity = await _db.Channels
+                .FirstOrDefaultAsync(c => c.PublicId == publicId);
+
+            return entity == null ? null : entity.Adapt<ChannelDto>();
+        }
+
+        public async Task<bool> DeleteChannelAsync(Guid id, Guid ownerId)
         {
             var entity = await _db.Channels.FirstOrDefaultAsync(c => c.Id == id);
 
@@ -115,6 +164,8 @@ namespace Infrastructure.Services.Channels
                 return false;
             }
 
+            if (entity.OwnerId != ownerId) throw new MemberNotAuthorizedException();
+
             _db.Channels.Remove(entity);
             await _db.SaveChangesAsync();
 
@@ -122,6 +173,45 @@ namespace Infrastructure.Services.Channels
 
             return true;
         }
-    }
 
+        private async Task<List<Tag>> ResolveTagsAsync(IEnumerable<string> tagNames)
+        {
+            var normalized = tagNames
+                .Select(t => t.ToLowerInvariant())
+                .Distinct()
+                .ToList();
+
+            var existing = await _db.Tags
+                .Where(t => normalized.Contains(t.NormalizedName))
+                .ToListAsync();
+
+            var existingNames = existing
+                .Select(t => t.NormalizedName)
+                .ToHashSet();
+
+            var newTags = normalized
+                .Where(n => !existingNames.Contains(n))
+                .Select(n => new Tag(n))
+                .ToList();
+
+            if (newTags.Any())
+                _db.Tags.AddRange(newTags);
+
+            return existing.Concat(newTags).ToList();
+        }
+
+        private async Task<int> GeneratePublicIdAsync()
+        {
+            var random = new Random();
+            int number;
+
+            do
+            {
+                number = random.Next(100000, 999999);
+            }
+            while (await _db.Channels.AnyAsync(c => c.PublicId == number));
+
+            return number;
+        }
+    }
 }
